requirements.txt
# Core LangChain and LangGraph
langchain>=0.2.0
langgraph>=0.2.0
langchain-groq>=0.1.0
langchain-community>=0.2.0

# LLM Provider - Groq Only
groq>=0.4.0

# Vector Store
chromadb>=0.4.0

# Text Processing
tiktoken>=0.5.0
pdfplumber>=0.10.0
pypdf>=3.0.0

# Embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0

# UI
streamlit>=1.28.0
gradio>=4.0.0

# Evaluation
trulens-eval>=0.19.0
langsmith>=0.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0